Implement jar download & safe strategy
Basically the program downloads the jar on a request if a file already doesn't exist. Then it schedules a task to delete the jar in 15 minutes. In that 15 minutes the downloaded jar will be used if a request on that resource happens.

There's one more problem tho: when I tested that change on offline, it worked fine, however when I put it on a vps things start to not work very well (maybe of the directory for saving the jars or something?)